+++ src/main/java/com/kihong/learn/java/phoneshop/exception/ResourceNotFoundException.java
@@ -3,7 +3,7 @@ package com.kihong.learn.java.phoneshop.exception;
 import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundException extends ApiException {
-	public ResourceNotFoundException(String resourceName, Integer id) {
+	public ResourceNotFoundException(String resourceName, Long id) {
 		super(HttpStatus.NOT_FOUND, String.format("%s with id =%d not found", resourceName, id));
 
 	}
